rename the favorite file
