+++ backend/routers/routes/sport.js
@@ -9,7 +9,7 @@ const authorization = require("../middlewares/authorization");
 const sportRouter = express.Router();
 
 sportRouter.post("/sports", addSport);
-sportRouter.delete("/sports/:id",authentication ,authorization('Admin'), deleteSport);
+sportRouter.delete("/sports/:id",authentication ,authorization('1'), deleteSport);
 sportRouter.get("/sports", getAllSports);
 
 module.exports = sportRouter;
